refactor: 메인홈 필터부분 수정
